Moved the `l4lb` `NetConf` struct into its own package.
