+++ cmd/l4lb/l4lb.go
@@ -20,26 +20,15 @@ import (
 	"os"
 	"runtime"
 
+	"github.com/dcos/dcos-cni/pkg/l4lb"
 	"github.com/dcos/dcos-cni/pkg/minuteman"
 	"github.com/dcos/dcos-cni/pkg/spartan"
 
 	"github.com/containernetworking/cni/pkg/invoke"
 	"github.com/containernetworking/cni/pkg/ip"
 	"github.com/containernetworking/cni/pkg/skel"
-	"github.com/containernetworking/cni/pkg/types"
 	"github.com/containernetworking/cni/pkg/version"
 )
-
-// We need this struct for de-duplicating the embedding of
-// `types.NetConf` and `minuteman.NetConf` in `NetConf`.
-type NetConf struct {
-	types.NetConf
-	Spartan   bool                   `json:"spartan, omitempty"`
-	Minuteman *minuteman.NetConf     `json:"minuteman, omitempty"`
-	Args      map[string]interface{} `json:"args, omitempty"`
-	MTU       int                    `json:"mtu, omitempty"`
-	Delegate  map[string]interface{} `json:"delegate, omitempty"`
-}
 
 // By default Spartan and Minuteman are specified to be enabled.
 
@@ -50,44 +39,8 @@ func init() {
 	runtime.LockOSThread()
 }
 
-func initConf() *NetConf {
-	conf := &NetConf{
-		Spartan: true,
-		Minuteman: &minuteman.NetConf{
-			Enable: true,
-		},
-	}
-
-	return conf
-}
-
-func setupDelegateConf(conf *NetConf) (delegateConf []byte, delegatePlugin string, err error) {
-	conf.Delegate["name"] = conf.Name
-	conf.Delegate["cniVersion"] = conf.CNIVersion
-	conf.Delegate["args"] = conf.Args
-
-	delegateConf, err = json.Marshal(conf.Delegate)
-	if err != nil {
-		err = fmt.Errorf("failed to marshall the delegate configuration: %s", err)
-		return
-	}
-
-	_, ok := conf.Delegate["type"]
-	if !ok {
-		err = fmt.Errorf("type field missing in delegate network: %s", conf.Delegate["name"])
-		return
-	}
-
-	delegatePlugin, ok = conf.Delegate["type"].(string)
-	if !ok {
-		err = fmt.Errorf("type field in delegate network %s has incorrect type, expected a `string`", conf.Delegate["name"])
-	}
-
-	return
-}
-
 func cmdAdd(args *skel.CmdArgs) error {
-	conf := initConf()
+	conf := l4lb.NewNetConf()
 
 	if err := json.Unmarshal(args.StdinData, conf); err != nil {
 		return fmt.Errorf("failed to load netconf: %s", err)
@@ -97,7 +50,7 @@ func cmdAdd(args *skel.CmdArgs) error {
 		return fmt.Errorf("failed to enable forwarding: %s", err)
 	}
 
-	delegateConf, delegatePlugin, err := setupDelegateConf(conf)
+	delegateConf, delegatePlugin, err := conf.SetupDelegateConf()
 	if err != nil {
 		return fmt.Errorf("failed to retrieve delegate configuration: %s", err)
 	}
@@ -148,7 +101,7 @@ func cmdAdd(args *skel.CmdArgs) error {
 }
 
 func cmdDel(args *skel.CmdArgs) error {
-	conf := initConf()
+	conf := l4lb.NewNetConf()
 
 	if err := json.Unmarshal(args.StdinData, conf); err != nil {
 		return fmt.Errorf("failed to load netconf: %s", err)
@@ -181,7 +134,7 @@ func cmdDel(args *skel.CmdArgs) error {
 	}
 
 	// Invoke the delegate plugin.
-	delegateConf, delegatePlugin, err := setupDelegateConf(conf)
+	delegateConf, delegatePlugin, err := conf.SetupDelegateConf()
 	if err != nil {
 		return fmt.Errorf("failed to retrieve delegate configuration: %s", err)
 	}
